complete rewrite of midi device
Added the concept of *MidiActions*.
It comprises is not only the action to be executed, when the MidiAction is triggered, but also the storing/saving of the binding of midi-messages in the device.
A *MidiAction* is a decorator that can be applied to Methods anywhere in the Program.
The method which is decorated contains the actual program-code that should be executed, when the *MidiAction* is triggered. In the most simple case it would be a method that is also a handler for a gui-element of the program.
the decorator registers the *MidiAction* in a global place. every time a midi action is used from some place in the code, it will tell the midi action which device should be used. the devace - just as before - is merely a container for the data that can be de/serialized easily

There are three types of midi actions:
  * MidiAction         - regular single-trigger midi action
  * MidiRowAction  - a midi action that stores a one-dimensional list of bindings
  * MidiGridAction   - a midi action that stores a two-dimensional list of bindings

Each of them comes in two variants:
   * normal: messages with velocity 0 will be ignored
   * continuous: all messages are interpreted

When midi-messages are learned, not the complete message, but only a key is stored in the device.
the key is the original midi messages without the velocity.
when the action is triggered, however, the velocity is simply routed through to the method.

This restriction makes it impossible to bind midi messages with the same signature that differ in velocity only to different midi actions. For instance a button that would send the same midi message with velocity 0 and 127 cannot trigger different things on push and release. However, this should not be encouraged anyways.
If this should ever become an issue in the future - say someone wants to add actions that have a relevant "on" and "off" part to them, it is recommended to subclass MidiAction and make a new model that reflects this behaviour (store two midi messages).