Fix loading of mono files
If soundfile is mono then soundfile library return an *one dimension* array
with data. With stereo or multichannel files it always return an array
containing one array per channel. So in order to have unified data structure we
need to put array of data for mono file in a an array to have an array with one
array for mono channel.